added documentation, fixed chat bubbles text rendering, moved Player class to its own file
