+++ client/js/Scenes/BootScene.js
@@ -1,5 +1,6 @@
 // import socket from './../socket.js';
 
+// eslint-disable-next-line no-undef
 class BootScene extends Phaser.Scene
 {
     constructor () {
