--- client/js/phasertest.js
@@ -1,4 +1,0 @@
-import socket from './socket.js';
-import BootScene from './Scenes/BootScene.js'
-
-
